epam.recipe-main/Epam.Recipe.AuthorizationService/Epam.Recipe.AuthorizationService.API/Middlewares/ExceptionMiddleware.cs
using Newtonsoft.Json;
using System.Net;

namespace Epam.Recipe.AuthorizationService.API.Middlewares
{
    public class ExceptionMiddleware
    {
        private readonly RequestDelegate _next;
        public ExceptionMiddleware(RequestDelegate next)
        {
            _next = next;
        }
        public async Task InvokeAsync(HttpContext context)
        {
            try { await _next(context); }
            catch (Exception ex)
            {
                await HandleExceptionAsync(context, ex);
            }
        }
        private async Task HandleExceptionAsync(HttpContext context, Exception ex)
        {
            context.Response.StatusCode = (int)HttpStatusCode.InternalServerError;
            context.Response.ContentType = "application/json";
            await context.Response.WriteAsync(JsonConvert.SerializeObject(
                new
                {
                    timestamp = DateTime.UtcNow.ToString("yyyy-MM-ddTHH:mm:ssZ"),
                    path = context.Request.Path.Value,
                    StatusCode = context.Response.StatusCode,
                    Message = ex.Message
                }));
        }
    }
}
